moved method from .h to .cpp
