+++ cmake-build-debug/CMakeFiles/GEN_Labo05.dir/build.make
@@ -126,13 +126,28 @@ CMakeFiles/GEN_Labo05.dir/Bill.cpp.s: cmake_force
 	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --green "Compiling CXX source to assembly CMakeFiles/GEN_Labo05.dir/Bill.cpp.s"
 	C:\PROGRA~1\HASKEL~1\86E4AC~1.5\mingw\bin\G__~1.EXE $(CXX_DEFINES) $(CXX_INCLUDES) $(CXX_FLAGS) -S C:\Users\Pierre-Olivier\Documents\BA4\GEN\labo\lab05\GEN_Labo_05\Bill.cpp -o CMakeFiles\GEN_Labo05.dir\Bill.cpp.s
 
+CMakeFiles/GEN_Labo05.dir/Rental.cpp.obj: CMakeFiles/GEN_Labo05.dir/flags.make
+CMakeFiles/GEN_Labo05.dir/Rental.cpp.obj: CMakeFiles/GEN_Labo05.dir/includes_CXX.rsp
+CMakeFiles/GEN_Labo05.dir/Rental.cpp.obj: ../Rental.cpp
+	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --green --progress-dir=C:\Users\Pierre-Olivier\Documents\BA4\GEN\labo\lab05\GEN_Labo_05\cmake-build-debug\CMakeFiles --progress-num=$(CMAKE_PROGRESS_6) "Building CXX object CMakeFiles/GEN_Labo05.dir/Rental.cpp.obj"
+	C:\PROGRA~1\HASKEL~1\86E4AC~1.5\mingw\bin\G__~1.EXE  $(CXX_DEFINES) $(CXX_INCLUDES) $(CXX_FLAGS) -o CMakeFiles\GEN_Labo05.dir\Rental.cpp.obj -c C:\Users\Pierre-Olivier\Documents\BA4\GEN\labo\lab05\GEN_Labo_05\Rental.cpp
+
+CMakeFiles/GEN_Labo05.dir/Rental.cpp.i: cmake_force
+	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --green "Preprocessing CXX source to CMakeFiles/GEN_Labo05.dir/Rental.cpp.i"
+	C:\PROGRA~1\HASKEL~1\86E4AC~1.5\mingw\bin\G__~1.EXE $(CXX_DEFINES) $(CXX_INCLUDES) $(CXX_FLAGS) -E C:\Users\Pierre-Olivier\Documents\BA4\GEN\labo\lab05\GEN_Labo_05\Rental.cpp > CMakeFiles\GEN_Labo05.dir\Rental.cpp.i
+
+CMakeFiles/GEN_Labo05.dir/Rental.cpp.s: cmake_force
+	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --green "Compiling CXX source to assembly CMakeFiles/GEN_Labo05.dir/Rental.cpp.s"
+	C:\PROGRA~1\HASKEL~1\86E4AC~1.5\mingw\bin\G__~1.EXE $(CXX_DEFINES) $(CXX_INCLUDES) $(CXX_FLAGS) -S C:\Users\Pierre-Olivier\Documents\BA4\GEN\labo\lab05\GEN_Labo_05\Rental.cpp -o CMakeFiles\GEN_Labo05.dir\Rental.cpp.s
+
 # Object files for target GEN_Labo05
 GEN_Labo05_OBJECTS = \
 "CMakeFiles/GEN_Labo05.dir/main.cpp.obj" \
 "CMakeFiles/GEN_Labo05.dir/Customer.cpp.obj" \
 "CMakeFiles/GEN_Labo05.dir/Movie.cpp.obj" \
 "CMakeFiles/GEN_Labo05.dir/test/CustomerTest.cpp.obj" \
-"CMakeFiles/GEN_Labo05.dir/Bill.cpp.obj"
+"CMakeFiles/GEN_Labo05.dir/Bill.cpp.obj" \
+"CMakeFiles/GEN_Labo05.dir/Rental.cpp.obj"
 
 # External object files for target GEN_Labo05
 GEN_Labo05_EXTERNAL_OBJECTS =
@@ -142,6 +157,7 @@ GEN_Labo05.exe: CMakeFiles/GEN_Labo05.dir/Customer.cpp.obj
 GEN_Labo05.exe: CMakeFiles/GEN_Labo05.dir/Movie.cpp.obj
 GEN_Labo05.exe: CMakeFiles/GEN_Labo05.dir/test/CustomerTest.cpp.obj
 GEN_Labo05.exe: CMakeFiles/GEN_Labo05.dir/Bill.cpp.obj
+GEN_Labo05.exe: CMakeFiles/GEN_Labo05.dir/Rental.cpp.obj
 GEN_Labo05.exe: CMakeFiles/GEN_Labo05.dir/build.make
 GEN_Labo05.exe: lib/libgtestd.a
 GEN_Labo05.exe: lib/libgtest_maind.a
@@ -149,7 +165,7 @@ GEN_Labo05.exe: lib/libgtestd.a
 GEN_Labo05.exe: CMakeFiles/GEN_Labo05.dir/linklibs.rsp
 GEN_Labo05.exe: CMakeFiles/GEN_Labo05.dir/objects1.rsp
 GEN_Labo05.exe: CMakeFiles/GEN_Labo05.dir/link.txt
-	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --green --bold --progress-dir=C:\Users\Pierre-Olivier\Documents\BA4\GEN\labo\lab05\GEN_Labo_05\cmake-build-debug\CMakeFiles --progress-num=$(CMAKE_PROGRESS_6) "Linking CXX executable GEN_Labo05.exe"
+	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --green --bold --progress-dir=C:\Users\Pierre-Olivier\Documents\BA4\GEN\labo\lab05\GEN_Labo_05\cmake-build-debug\CMakeFiles --progress-num=$(CMAKE_PROGRESS_7) "Linking CXX executable GEN_Labo05.exe"
 	$(CMAKE_COMMAND) -E cmake_link_script CMakeFiles\GEN_Labo05.dir\link.txt --verbose=$(VERBOSE)
 
 # Rule to build all files generated by this target.
