+++ cmake-build-debug/CMakeFiles/GEN_Labo05.dir/cmake_clean.cmake
@@ -2,6 +2,7 @@ file(REMOVE_RECURSE
   "CMakeFiles/GEN_Labo05.dir/Bill.cpp.obj"
   "CMakeFiles/GEN_Labo05.dir/Customer.cpp.obj"
   "CMakeFiles/GEN_Labo05.dir/Movie.cpp.obj"
+  "CMakeFiles/GEN_Labo05.dir/Rental.cpp.obj"
   "CMakeFiles/GEN_Labo05.dir/main.cpp.obj"
   "CMakeFiles/GEN_Labo05.dir/test/CustomerTest.cpp.obj"
   "GEN_Labo05.exe"
