+++ cmake-build-debug/CMakeFiles/Makefile2
@@ -132,16 +132,16 @@ test/googletest/clean: test/googletest/CMakeFiles/gtest_main.dir/clean
 # Target rules for target CMakeFiles/GEN_Labo05.dir
 
 # All Build rule for target.
-CMakeFiles/GEN_Labo05.dir/all: test/googletest/CMakeFiles/gtest.dir/all
 CMakeFiles/GEN_Labo05.dir/all: test/googletest/CMakeFiles/gtest_main.dir/all
+CMakeFiles/GEN_Labo05.dir/all: test/googletest/CMakeFiles/gtest.dir/all
 	$(MAKE) -f CMakeFiles\GEN_Labo05.dir\build.make CMakeFiles/GEN_Labo05.dir/depend
 	$(MAKE) -f CMakeFiles\GEN_Labo05.dir\build.make CMakeFiles/GEN_Labo05.dir/build
-	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --progress-dir=C:\Users\Pierre-Olivier\Documents\BA4\GEN\labo\lab05\GEN_Labo_05\cmake-build-debug\CMakeFiles --progress-num=1,2,3,4,5,6 "Built target GEN_Labo05"
+	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --progress-dir=C:\Users\Pierre-Olivier\Documents\BA4\GEN\labo\lab05\GEN_Labo_05\cmake-build-debug\CMakeFiles --progress-num=1,2,3,4,5,6,7 "Built target GEN_Labo05"
 .PHONY : CMakeFiles/GEN_Labo05.dir/all
 
 # Build rule for subdir invocation for target.
 CMakeFiles/GEN_Labo05.dir/rule: cmake_check_build_system
-	$(CMAKE_COMMAND) -E cmake_progress_start C:\Users\Pierre-Olivier\Documents\BA4\GEN\labo\lab05\GEN_Labo_05\cmake-build-debug\CMakeFiles 10
+	$(CMAKE_COMMAND) -E cmake_progress_start C:\Users\Pierre-Olivier\Documents\BA4\GEN\labo\lab05\GEN_Labo_05\cmake-build-debug\CMakeFiles 11
 	$(MAKE) -f CMakeFiles\Makefile2 CMakeFiles/GEN_Labo05.dir/all
 	$(CMAKE_COMMAND) -E cmake_progress_start C:\Users\Pierre-Olivier\Documents\BA4\GEN\labo\lab05\GEN_Labo_05\cmake-build-debug\CMakeFiles 0
 .PHONY : CMakeFiles/GEN_Labo05.dir/rule
@@ -160,11 +160,11 @@ CMakeFiles/GEN_Labo05.dir/clean:
 # Target rules for target test/googlemock/CMakeFiles/gmock_main.dir
 
 # All Build rule for target.
-test/googlemock/CMakeFiles/gmock_main.dir/all: test/googlemock/CMakeFiles/gmock.dir/all
 test/googlemock/CMakeFiles/gmock_main.dir/all: test/googletest/CMakeFiles/gtest.dir/all
+test/googlemock/CMakeFiles/gmock_main.dir/all: test/googlemock/CMakeFiles/gmock.dir/all
 	$(MAKE) -f test\googlemock\CMakeFiles\gmock_main.dir\build.make test/googlemock/CMakeFiles/gmock_main.dir/depend
 	$(MAKE) -f test\googlemock\CMakeFiles\gmock_main.dir\build.make test/googlemock/CMakeFiles/gmock_main.dir/build
-	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --progress-dir=C:\Users\Pierre-Olivier\Documents\BA4\GEN\labo\lab05\GEN_Labo_05\cmake-build-debug\CMakeFiles --progress-num=9,10 "Built target gmock_main"
+	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --progress-dir=C:\Users\Pierre-Olivier\Documents\BA4\GEN\labo\lab05\GEN_Labo_05\cmake-build-debug\CMakeFiles --progress-num=10,11 "Built target gmock_main"
 .PHONY : test/googlemock/CMakeFiles/gmock_main.dir/all
 
 # Build rule for subdir invocation for target.
@@ -191,7 +191,7 @@ test/googlemock/CMakeFiles/gmock_main.dir/clean:
 test/googlemock/CMakeFiles/gmock.dir/all: test/googletest/CMakeFiles/gtest.dir/all
 	$(MAKE) -f test\googlemock\CMakeFiles\gmock.dir\build.make test/googlemock/CMakeFiles/gmock.dir/depend
 	$(MAKE) -f test\googlemock\CMakeFiles\gmock.dir\build.make test/googlemock/CMakeFiles/gmock.dir/build
-	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --progress-dir=C:\Users\Pierre-Olivier\Documents\BA4\GEN\labo\lab05\GEN_Labo_05\cmake-build-debug\CMakeFiles --progress-num=7,8 "Built target gmock"
+	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --progress-dir=C:\Users\Pierre-Olivier\Documents\BA4\GEN\labo\lab05\GEN_Labo_05\cmake-build-debug\CMakeFiles --progress-num=8,9 "Built target gmock"
 .PHONY : test/googlemock/CMakeFiles/gmock.dir/all
 
 # Build rule for subdir invocation for target.
@@ -218,7 +218,7 @@ test/googlemock/CMakeFiles/gmock.dir/clean:
 test/googletest/CMakeFiles/gtest.dir/all:
 	$(MAKE) -f test\googletest\CMakeFiles\gtest.dir\build.make test/googletest/CMakeFiles/gtest.dir/depend
 	$(MAKE) -f test\googletest\CMakeFiles\gtest.dir\build.make test/googletest/CMakeFiles/gtest.dir/build
-	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --progress-dir=C:\Users\Pierre-Olivier\Documents\BA4\GEN\labo\lab05\GEN_Labo_05\cmake-build-debug\CMakeFiles --progress-num=11,12 "Built target gtest"
+	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --progress-dir=C:\Users\Pierre-Olivier\Documents\BA4\GEN\labo\lab05\GEN_Labo_05\cmake-build-debug\CMakeFiles --progress-num=12,13 "Built target gtest"
 .PHONY : test/googletest/CMakeFiles/gtest.dir/all
 
 # Build rule for subdir invocation for target.
@@ -245,7 +245,7 @@ test/googletest/CMakeFiles/gtest.dir/clean:
 test/googletest/CMakeFiles/gtest_main.dir/all: test/googletest/CMakeFiles/gtest.dir/all
 	$(MAKE) -f test\googletest\CMakeFiles\gtest_main.dir\build.make test/googletest/CMakeFiles/gtest_main.dir/depend
 	$(MAKE) -f test\googletest\CMakeFiles\gtest_main.dir\build.make test/googletest/CMakeFiles/gtest_main.dir/build
-	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --progress-dir=C:\Users\Pierre-Olivier\Documents\BA4\GEN\labo\lab05\GEN_Labo_05\cmake-build-debug\CMakeFiles --progress-num=13,14 "Built target gtest_main"
+	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --progress-dir=C:\Users\Pierre-Olivier\Documents\BA4\GEN\labo\lab05\GEN_Labo_05\cmake-build-debug\CMakeFiles --progress-num=14,15 "Built target gtest_main"
 .PHONY : test/googletest/CMakeFiles/gtest_main.dir/all
 
 # Build rule for subdir invocation for target.
