+++ cmake-build-debug/test/googlemock/CMakeFiles/gmock.dir/progress.make
@@ -1,3 +1,3 @@
-CMAKE_PROGRESS_1 = 7
-CMAKE_PROGRESS_2 = 8
+CMAKE_PROGRESS_1 = 8
+CMAKE_PROGRESS_2 = 9
