+++ cmake-build-debug/test/googlemock/CMakeFiles/gmock_main.dir/progress.make
@@ -1,3 +1,3 @@
-CMAKE_PROGRESS_1 = 9
-CMAKE_PROGRESS_2 = 10
+CMAKE_PROGRESS_1 = 10
+CMAKE_PROGRESS_2 = 11
